AST: Refactor FeatureSet out into its own file.
NFC.
