+++ lib/AST/ASTPrinter.cpp
@@ -14,6 +14,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "swift/AST/ASTPrinter.h"
+#include "FeatureSet.h"
 #include "InlinableText.h"
 #include "swift/AST/ASTContext.h"
 #include "swift/AST/ASTMangler.h"
@@ -23,7 +24,6 @@
 #include "swift/AST/ClangModuleLoader.h"
 #include "swift/AST/Comment.h"
 #include "swift/AST/Decl.h"
-#include "swift/AST/ExistentialLayout.h"
 #include "swift/AST/Expr.h"
 #include "swift/AST/FileUnit.h"
 #include "swift/AST/GenericEnvironment.h"
@@ -45,7 +45,6 @@
 #include "swift/AST/Types.h"
 #include "swift/Basic/Defer.h"
 #include "swift/Basic/Feature.h"
-#include "swift/Basic/FixedBitSet.h"
 #include "swift/Basic/PrimitiveParsing.h"
 #include "swift/Basic/QuotedString.h"
 #include "swift/Basic/STLExtras.h"
@@ -3011,236 +3010,12 @@ void PrintAST::printExtension(ExtensionDecl *decl) {
   }
 }
 
-/// Functions to determine which features a particular declaration uses. The
-/// usesFeatureNNN functions correspond to the features in Features.def.
-
-static bool usesFeatureStaticAssert(Decl *decl) {
-  return false;
-}
-
-#define BASELINE_LANGUAGE_FEATURE(FeatureName, SENumber, Description) \
-static bool usesFeature##FeatureName(Decl *decl) { \
-  return false; \
-}
-#define LANGUAGE_FEATURE(FeatureName, SENumber, Description)
-#include "swift/Basic/Features.def"
-
-static bool usesFeatureMacros(Decl *decl) {
-  return isa<MacroDecl>(decl);
-}
-
-static bool usesFeatureFreestandingMacros(Decl *decl) {
-  auto macro = dyn_cast<MacroDecl>(decl);
-  if (!macro)
-    return false;
-
-  return macro->getMacroRoles().contains(MacroRole::Declaration);
-}
-
-static bool usesFeatureExpressionMacroDefaultArguments(Decl *decl) {
-  if (auto func = dyn_cast<AbstractFunctionDecl>(decl)) {
-    for (auto param : *func->getParameters()) {
-      if (param->getDefaultArgumentKind() ==
-          DefaultArgumentKind::ExpressionMacro)
-        return true;
-    }
-  }
-
-  return false;
-}
-
-static bool usesFeatureBuiltinStoreRaw(Decl *decl) { return false; }
-
-static bool usesFeatureCodeItemMacros(Decl *decl) {
-  auto macro = dyn_cast<MacroDecl>(decl);
-  if (!macro)
-    return false;
-
-  return macro->getMacroRoles().contains(MacroRole::CodeItem);
-}
-
-static bool usesFeatureExtensionMacros(Decl *decl) {
-  auto macro = dyn_cast<MacroDecl>(decl);
-  if (!macro)
-    return false;
-
-  return macro->getMacroRoles().contains(MacroRole::Extension);
-}
-
-static bool usesFeatureBodyMacros(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureExtensionMacroAttr(Decl *decl) {
-  return usesFeatureExtensionMacros(decl);
-}
-
 static void suppressingFeatureExtensionMacroAttr(PrintOptions &options,
                                                  llvm::function_ref<void()> action) {
   bool originalPrintExtensionMacroAttrs = options.PrintExtensionMacroAttributes;
   options.PrintExtensionMacroAttributes = false;
   action();
   options.PrintExtensionMacroAttributes = originalPrintExtensionMacroAttrs;
-}
-
-static bool usesFeatureAttachedMacros(Decl *decl) {
-  auto macro = dyn_cast<MacroDecl>(decl);
-  if (!macro)
-    return false;
-
-  return static_cast<bool>(macro->getMacroRoles() & getAttachedMacroRoles());
-}
-
-static bool usesFeatureRethrowsProtocol(
-    Decl *decl, SmallPtrSet<Decl *, 16> &checked) {
-  // Make sure we don't recurse.
-  if (!checked.insert(decl).second)
-    return false;
-
-  // Check an inheritance clause for a marker protocol.
-  auto checkInherited = [&](InheritedTypes inherited) -> bool {
-    for (unsigned i : inherited.getIndices()) {
-      if (auto inheritedType = inherited.getResolvedType(i)) {
-        if (inheritedType->isExistentialType()) {
-          auto layout = inheritedType->getExistentialLayout();
-          for (ProtocolDecl *proto : layout.getProtocols()) {
-            if (usesFeatureRethrowsProtocol(proto, checked))
-              return true;
-          }
-        }
-      }
-    }
-
-    return false;
-  };
-
-  if (auto nominal = dyn_cast<NominalTypeDecl>(decl)) {
-    if (checkInherited(nominal->getInherited()))
-      return true;
-  }
-
-  if (auto proto = dyn_cast<ProtocolDecl>(decl)) {
-    if (proto->getAttrs().hasAttribute<AtRethrowsAttr>())
-      return true;
-  }
-
-  if (auto ext = dyn_cast<ExtensionDecl>(decl)) {
-    if (auto nominal = ext->getSelfNominalTypeDecl())
-      if (usesFeatureRethrowsProtocol(nominal, checked))
-        return true;
-
-    if (checkInherited(ext->getInherited()))
-      return true;
-  }
-
-  if (auto genericSig = decl->getInnermostDeclContext()
-          ->getGenericSignatureOfContext()) {
-    for (const auto &req : genericSig.getRequirements()) {
-      if (req.getKind() == RequirementKind::Conformance &&
-          usesFeatureRethrowsProtocol(req.getProtocolDecl(), checked))
-        return true;
-    }
-  }
-
-  if (auto value = dyn_cast<ValueDecl>(decl)) {
-    if (Type type = value->getInterfaceType()) {
-      bool hasRethrowsProtocol = type.findIf([&](Type type) {
-        if (auto nominal = type->getAnyNominal()) {
-          if (usesFeatureRethrowsProtocol(nominal, checked))
-            return true;
-        }
-
-        return false;
-      });
-
-      if (hasRethrowsProtocol)
-        return true;
-    }
-  }
-
-  return false;
-}
-
-static bool usesFeatureRethrowsProtocol(Decl *decl) {
-  SmallPtrSet<Decl *, 16> checked;
-  return usesFeatureRethrowsProtocol(decl, checked);
-}
-
-static bool usesFeatureRetroactiveAttribute(Decl *decl) {
-  auto ext = dyn_cast<ExtensionDecl>(decl);
-  if (!ext)
-    return false;
-
-  return llvm::any_of(
-      ext->getInherited().getEntries(),
-      [](const InheritedEntry &entry) { return entry.isRetroactive(); });
-}
-
-/// Does the interface of this declaration use a type for which the
-/// given predicate returns true?
-static bool usesTypeMatching(Decl *decl, llvm::function_ref<bool(Type)> fn) {
-  if (auto value = dyn_cast<ValueDecl>(decl)) {
-    if (Type type = value->getInterfaceType()) {
-      return type.findIf(fn);
-    }
-  }
-
-  return false;
-}
-
-static bool usesFeatureBuiltinBuildTaskExecutorRef(Decl *decl) { return false; }
-
-static bool usesFeatureBuiltinBuildExecutor(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureBuiltinBuildComplexEqualityExecutor(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureBuiltinCreateAsyncTaskInGroupWithExecutor(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureBuiltinCreateAsyncDiscardingTaskInGroup(Decl *decl) {
-  return false;
-}
-
-static bool
-usesFeatureBuiltinCreateAsyncDiscardingTaskInGroupWithExecutor(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureBuiltinCopy(Decl *decl) { return false; }
-
-static bool usesFeatureBuiltinTaskRunInline(Decl *) { return false; }
-
-static bool usesFeatureBuiltinUnprotectedAddressOf(Decl *) { return false; }
-
-static bool usesFeatureBuiltinCreateTaskGroupWithFlags(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureSpecializeAttributeWithAvailability(Decl *decl) {
-  if (auto func = dyn_cast<AbstractFunctionDecl>(decl)) {
-    for (auto specialize : func->getAttrs().getAttributes<SpecializeAttr>()) {
-      if (!specialize->getAvailableAttrs().empty())
-        return true;
-    }
-  }
-  return false;
-}
-
-static bool usesFeatureParserRoundTrip(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureParserValidation(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureParserDiagnostics(Decl *decl) {
-  return false;
 }
 
 static void suppressingFeatureSpecializeAttributeWithAvailability(
@@ -3251,30 +3026,6 @@ static void suppressingFeatureSpecializeAttributeWithAvailability(
   action();
 }
 
-static bool usesFeatureBuiltinStackAlloc(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureEmbedded(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureBuiltinUnprotectedStackAlloc(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureBuiltinAllocVector(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureBuiltinAssumeAlignment(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureUnsafeInheritExecutor(Decl *decl) {
-  return decl->getAttrs().hasAttribute<UnsafeInheritExecutorAttr>();
-}
-
 static void suppressingFeatureUnsafeInheritExecutor(PrintOptions &options,
                                         llvm::function_ref<void()> action) {
   unsigned originalExcludeAttrCount = options.ExcludeAttrList.size();
@@ -3283,29 +3034,12 @@ static void suppressingFeatureUnsafeInheritExecutor(PrintOptions &options,
   options.ExcludeAttrList.resize(originalExcludeAttrCount);
 }
 
-static bool usesFeaturePrimaryAssociatedTypes2(Decl *decl) {
-  if (auto *protoDecl = dyn_cast<ProtocolDecl>(decl)) {
-    if (protoDecl->getPrimaryAssociatedTypes().size() > 0)
-      return true;
-  }
-
-  return false;
-}
-
-static bool usesFeatureSendableCompletionHandlers(Decl *decl) {
-  return false;
-}
-
 static void suppressingFeaturePrimaryAssociatedTypes2(PrintOptions &options,
                                          llvm::function_ref<void()> action) {
   bool originalPrintPrimaryAssociatedTypes = options.PrintPrimaryAssociatedTypes;
   options.PrintPrimaryAssociatedTypes = false;
   action();
   options.PrintPrimaryAssociatedTypes = originalPrintPrimaryAssociatedTypes;
-}
-
-static bool usesFeatureUnavailableFromAsync(Decl *decl) {
-  return decl->getAttrs().hasAttribute<UnavailableFromAsyncAttr>();
 }
 
 static void
@@ -3317,15 +3051,6 @@ suppressingFeatureUnavailableFromAsync(PrintOptions &options,
   options.ExcludeAttrList.resize(originalExcludeAttrCount);
 }
 
-static bool usesFeatureNoAsyncAvailability(Decl *decl) {
-   return decl->getAttrs().getNoAsync(decl->getASTContext()) != nullptr;
-}
-
-static bool usesFeatureAssociatedTypeAvailability(Decl *decl) {
-  return isa<AssociatedTypeDecl>(decl) &&
-      decl->getAttrs().hasAttribute<AvailableAttr>();
-}
-
 static void
 suppressingFeatureAssociatedTypeAvailability(
     PrintOptions &options, llvm::function_ref<void()> action) {
@@ -3335,299 +3060,12 @@ suppressingFeatureAssociatedTypeAvailability(
   options.ExcludeAttrList.resize(originalExcludeAttrCount);
 }
 
-static bool usesFeatureAsyncSequenceFailure(Decl *decl) {
-  if (auto proto = dyn_cast<ProtocolDecl>(decl)) {
-    return isImplicitRethrowsProtocol(proto);
-  }
-
-  return false;
-}
-
 static void
 suppressingFeatureAsyncSequenceFailure(
     PrintOptions &options, llvm::function_ref<void()> action) {
   llvm::SaveAndRestore<bool> saved(options.AsyncSequenceRethrows, true);
   action();
 }
-
-static bool usesFeatureBuiltinIntLiteralAccessors(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureConciseMagicFile(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureExistentialAny(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureStrictConcurrency(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureDeprecateApplicationMain(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureGroupActorErrors(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureImportObjcForwardDeclarations(Decl *decl) {
-  ClangNode clangNode = decl->getClangNode();
-  if (!clangNode)
-    return false;
-
-  const clang::Decl *clangDecl = clangNode.getAsDecl();
-  if (!clangDecl)
-    return false;
-
-  if (auto objCInterfaceDecl = dyn_cast<clang::ObjCInterfaceDecl>(clangDecl))
-    return !objCInterfaceDecl->hasDefinition();
-
-  if (auto objCProtocolDecl = dyn_cast<clang::ObjCProtocolDecl>(clangDecl))
-    return !objCProtocolDecl->hasDefinition();
-
-  return false;
-}
-
-static bool usesFeatureImplicitSome(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureForwardTrailingClosures(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureBareSlashRegexLiterals(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureTupleConformances(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureSymbolLinkageMarkers(Decl *decl) {
-  auto &attrs = decl->getAttrs();
-  return std::any_of(attrs.begin(), attrs.end(), [](auto *attr) {
-    if (isa<UsedAttr>(attr))
-      return true;
-    if (isa<SectionAttr>(attr))
-      return true;
-    return false;
-  });
-}
-
-static bool usesFeatureLayoutPrespecialization(Decl *decl) {
-  auto &attrs = decl->getAttrs();
-  return std::any_of(attrs.begin(), attrs.end(), [](auto *attr) {
-    if (auto *specialize = dyn_cast<SpecializeAttr>(attr)) {
-      return !specialize->getTypeErasedParams().empty();
-    }
-    return false;
-  });
-}
-
-static bool usesFeatureLayoutStringValueWitnesses(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureLayoutStringValueWitnessesInstantiation(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureAccessLevelOnImport(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureNamedOpaqueTypes(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureFlowSensitiveConcurrencyCaptures(Decl *decl) {
-  return false;
-}
-
-/// \param isRelevantInverse the function used to inspect a mark corresponding
-/// to an inverse to determine whether it "has" an inverse that we care about.
-static bool hasInverse(
-    Decl *decl, InvertibleProtocolKind ip,
-    std::function<bool(InverseMarking::Mark const &)> isRelevantInverse) {
-
-  auto getTypeDecl = [](Type type) -> TypeDecl* {
-    if (auto genericTy = type->getAnyGeneric())
-      return genericTy;
-    if (auto gtpt = dyn_cast<GenericTypeParamType>(type))
-      return gtpt->getDecl();
-    return nullptr;
-  };
-
-  if (auto *extension = dyn_cast<ExtensionDecl>(decl)) {
-    if (auto *nominal = extension->getSelfNominalTypeDecl())
-      return hasInverse(nominal, ip, isRelevantInverse);
-    return false;
-  }
-
-  auto hasInverseInType = [&](Type type) {
-    return type.findIf([&](Type type) -> bool {
-      if (auto *typeDecl = getTypeDecl(type))
-        return hasInverse(typeDecl, ip, isRelevantInverse);
-      return false;
-    });
-  };
-
-  if (auto *TD = dyn_cast<TypeDecl>(decl)) {
-    if (auto *alias = dyn_cast<TypeAliasDecl>(TD))
-      return hasInverseInType(alias->getUnderlyingType());
-
-    if (auto *NTD = dyn_cast<NominalTypeDecl>(TD)) {
-      if (isRelevantInverse(NTD->hasInverseMarking(ip)))
-        return true;
-    }
-
-    if (auto *P = dyn_cast<ProtocolDecl>(TD)) {
-      // Check the protocol's associated types too.
-      return llvm::any_of(
-          P->getAssociatedTypeMembers(), [&](AssociatedTypeDecl *ATD) {
-            return isRelevantInverse(ATD->hasInverseMarking(ip));
-          });
-    }
-
-    return false;
-  }
-
-  if (auto *VD = dyn_cast<ValueDecl>(decl)) {
-    if (VD->hasInterfaceType())
-      return hasInverseInType(VD->getInterfaceType());
-  }
-
-  return false;
-}
-
-static bool usesFeatureMoveOnly(Decl *decl) {
-  return hasInverse(decl, InvertibleProtocolKind::Copyable,
-                    [](auto &marking) -> bool {
-                      return marking.is(InverseMarking::Kind::LegacyExplicit);
-                    });
-}
-
-static bool usesFeatureLazyImmediate(Decl *D) { return false; }
-
-static bool usesFeatureMoveOnlyClasses(Decl *decl) {
-  return isa<ClassDecl>(decl) && usesFeatureMoveOnly(decl);
-}
-
-static bool usesFeatureMoveOnlyTuples(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureNoImplicitCopy(Decl *decl) {
-  return decl->isNoImplicitCopy();
-}
-
-static bool usesFeatureOldOwnershipOperatorSpellings(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureMoveOnlyEnumDeinits(Decl *decl) {
-  if (auto *ei = dyn_cast<EnumDecl>(decl)) {
-    return usesFeatureMoveOnly(ei) && ei->getValueTypeDestructor();
-  }
-  return false;
-}
-
-static bool usesFeatureMoveOnlyResilientTypes(Decl *decl) {
-  if (auto *nomDecl = dyn_cast<NominalTypeDecl>(decl))
-    return nomDecl->isResilient() && usesFeatureMoveOnly(decl);
-  return false;
-}
-
-static bool usesFeatureMoveOnlyPartialConsumption(Decl *decl) {
-  // Partial consumption does not affect declarations directly.
-  return false;
-}
-
-static bool usesFeatureMoveOnlyPartialReinitialization(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureNoncopyableGenerics(Decl *decl) {
-  auto checkInverseMarking = [](auto &marking) -> bool {
-    switch (marking.getKind()) {
-    case InverseMarking::Kind::None:
-    case InverseMarking::Kind::LegacyExplicit: // covered by other checks.
-      return false;
-
-    case InverseMarking::Kind::Explicit:
-    case InverseMarking::Kind::Inferred:
-      return true;
-    }
-  };
-
-  return hasInverse(decl, InvertibleProtocolKind::Copyable,
-                    checkInverseMarking) ||
-         hasInverse(decl, InvertibleProtocolKind::Escapable,
-                    checkInverseMarking);
-}
-
-static bool usesFeatureOneWayClosureParameters(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureOpaqueTypeErasure(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureDifferentiableProgramming(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureForwardModeDifferentiation(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureAdditiveArithmeticDerivedConformances(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureParserASTGen(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureBuiltinMacros(Decl *decl) {
-  return false;
-}
-
-
-static bool usesFeatureDisableOutwardActorInference(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureInternalImportsByDefault(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureImportSymbolicCXXDecls(Decl *decl) { return false; }
-
-static bool usesFeatureGenerateBindingsForThrowingFunctionsInCXX(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureFreestandingExpressionMacros(Decl *decl) {
-  auto macro = dyn_cast<MacroDecl>(decl);
-  if (!macro)
-    return false;
-
-  return macro->getMacroRoles().contains(MacroRole::Expression);
-}
-
-static bool usesFeatureLexicalLifetimes(Decl *decl) {
-  return decl->getAttrs().hasAttribute<EagerMoveAttr>()
-         || decl->getAttrs().hasAttribute<NoEagerMoveAttr>()
-         || decl->getAttrs().hasAttribute<LexicalLifetimesAttr>();
-}
-
 static void
 suppressingFeatureLexicalLifetimes(PrintOptions &options,
                                    llvm::function_ref<void()> action) {
@@ -3655,256 +3093,12 @@ static void suppressingFeatureRetroactiveAttribute(
   action();
 }
 
-static bool usesFeatureReferenceBindings(Decl *decl) {
-  auto *vd = dyn_cast<VarDecl>(decl);
-  return vd && vd->getIntroducer() == VarDecl::Introducer::InOut;
-}
-
-static bool usesFeatureBuiltinModule(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureRawLayout(Decl *decl) {
-  return decl->getAttrs().hasAttribute<RawLayoutAttr>();
-}
-
-static bool usesFeatureStructLetDestructuring(Decl *decl) {
-  auto sd = dyn_cast<StructDecl>(decl);
-  if (!sd)
-    return false;
-    
-  for (auto member : sd->getStoredProperties()) {
-    if (!member->isLet())
-      continue;
-    
-    auto init = member->getParentPattern();
-    if (!init)
-      continue;
-      
-    if (!init->getSingleVar())
-      return true;
-  }
-  
-  return false;
-}
-
-static bool usesFeatureNonescapableTypes(Decl *decl) {
-  if (decl->getAttrs().hasAttribute<NonEscapableAttr>() ||
-      decl->getAttrs().hasAttribute<UnsafeNonEscapableResultAttr>()) {
-    return true;
-  }
-  auto *fd = dyn_cast<FuncDecl>(decl);
-  if (fd && fd->getAttrs().getAttribute(DeclAttrKind::ResultDependsOnSelf)) {
-    return true;
-  }
-  auto *pd = dyn_cast<ParamDecl>(decl);
-  if (pd && pd->hasResultDependsOn()) {
-    return true;
-  }
-  return false;
-}
-
-static bool usesFeatureFixedArrays(Decl *decl) {
-  return false;
-}
-
-static bool hasParameterPacks(Decl *decl) {
-  if (auto genericContext = decl->getAsGenericContext()) {
-    auto sig = genericContext->getGenericSignature();
-    if (llvm::any_of(
-          sig.getGenericParams(),
-          [&](const GenericTypeParamType *GP) { return GP->isParameterPack(); })) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-/// A declaration needs the $ParameterPacks feature if it declares a
-/// generic parameter pack, or if its type references a generic nominal
-/// or type alias which declares a generic parameter pack.
-static bool usesFeatureParameterPacks(Decl *decl) {
-  if (hasParameterPacks(decl))
-    return true;
-
-  if (auto *valueDecl = dyn_cast<ValueDecl>(decl)) {
-    if (valueDecl->getInterfaceType().findIf(
-        [&](Type t) {
-          if (auto *alias = dyn_cast<TypeAliasType>(t.getPointer()))
-            return hasParameterPacks(alias->getDecl());
-          if (auto *nominal = t->getAnyNominal())
-            return hasParameterPacks(nominal);
-
-          return false;
-        })) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-static bool usesFeatureRegionBasedIsolation(Decl *decl) { return false; }
-
-static bool usesFeatureGlobalConcurrency(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureIsolatedDefaultValues(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureInferSendableFromCaptures(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureOptionalIsolatedParameters(Decl *decl) {
-  auto *value = dyn_cast<ValueDecl>(decl);
-  if (!value)
-    return false;
-
-  auto *paramList = getParameterList(value);
-  if (!paramList)
-    return false;
-
-  for (auto param : *paramList) {
-    if (param->isIsolated()) {
-      auto paramType = param->getInterfaceType();
-      return !paramType->getOptionalObjectType().isNull();
-    }
-  }
-
-  return false;
-}
-
-static bool usesFeaturePlaygroundExtendedCallbacks(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureThenStatements(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureDoExpressions(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureImplicitLastExprResults(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureNewCxxMethodSafetyHeuristics(Decl *decl) {
-  return decl->hasClangNode();
-}
-
-static bool usesFeatureFullTypedThrows(Decl *decl) {
-  return false;
-}
-
-static bool usesFeatureTypedThrows(Decl *decl) {
-  if (auto func = dyn_cast<AbstractFunctionDecl>(decl)) {
-    struct Walker : public TypeWalker {
-      bool hasTypedThrows = false;
-
-      Action walkToTypePre(Type ty) override {
-        if (auto funcType = ty->getAs<AnyFunctionType>()) {
-          if (funcType->hasThrownError()) {
-            hasTypedThrows = true;
-            return Action::Stop;
-          }
-        }
-
-        return Action::Continue;
-      }
-    };
-
-    Walker walker;
-    func->getInterfaceType().walk(walker);
-    return walker.hasTypedThrows;
-  }
-
-  return false;
-}
-
-static bool usesFeatureExtern(Decl *decl) {
-  return decl->getAttrs().hasAttribute<ExternAttr>();
-}
-
 static void suppressingFeatureExtern(PrintOptions &options,
                                      llvm::function_ref<void()> action) {
   unsigned originalExcludeAttrCount = options.ExcludeAttrList.size();
   options.ExcludeAttrList.push_back(DeclAttrKind::Extern);
   action();
   options.ExcludeAttrList.resize(originalExcludeAttrCount);
-}
-
-static bool usesFeatureStaticExclusiveOnly(Decl *decl) {
-  return decl->getAttrs().hasAttribute<StaticExclusiveOnlyAttr>();
-}
-
-static bool usesFeatureExtractConstantsFromMembers(Decl *decl) {
-  return decl->getAttrs().hasAttribute<ExtractConstantsFromMembersAttr>();
-}
-
-static bool usesFeatureBitwiseCopyable(Decl *decl) { return false; }
-
-static bool usesFeatureTransferringArgsAndResults(Decl *decl) {
-  if (auto *pd = dyn_cast<ParamDecl>(decl))
-    if (pd->isTransferring())
-      return true;
-
-  if (auto *fDecl = dyn_cast<FuncDecl>(decl)) {
-    auto fnTy = fDecl->getInterfaceType();
-    bool hasTransferring = false;
-    if (auto *ft = llvm::dyn_cast_if_present<FunctionType>(fnTy)) {
-      if (ft->hasExtInfo())
-        hasTransferring = ft->hasTransferringResult();
-    } else if (auto *ft =
-               llvm::dyn_cast_if_present<GenericFunctionType>(fnTy)) {
-      if (ft->hasExtInfo())
-        hasTransferring = ft->hasTransferringResult();
-    }
-    if (hasTransferring)
-      return true;
-  }
-
-  return false;
-}
-
-static bool usesFeatureDynamicActorIsolation(Decl *decl) {
-  auto usesPreconcurrencyConformance = [&](const InheritedTypes &inherited) {
-    return llvm::any_of(
-        inherited.getEntries(),
-        [](const InheritedEntry &entry) { return entry.isPreconcurrency(); });
-  };
-
-  if (auto *T = dyn_cast<TypeDecl>(decl))
-    return usesPreconcurrencyConformance(T->getInherited());
-
-  if (auto *E = dyn_cast<ExtensionDecl>(decl)) {
-    // If type has `@preconcurrency` conformance(s) all of its
-    // extensions have to be guarded by the flag too.
-    if (auto *T = dyn_cast<TypeDecl>(E->getExtendedNominal())) {
-      if (usesPreconcurrencyConformance(T->getInherited()))
-        return true;
-    }
-
-    return usesPreconcurrencyConformance(E->getInherited());
-  }
-
-  return false;
-}
-
-static bool usesFeatureBorrowingSwitch(Decl *decl) { return false; }
-
-static bool usesFeatureIsolatedAny(Decl *decl) {
-  return usesTypeMatching(decl, [](Type type) {
-    if (auto fnType = type->getAs<AnyFunctionType>()) {
-      return fnType->getIsolation().isErased();
-    }
-    return false;
-  });
 }
 
 /// Suppress the printing of a particular feature.
@@ -3921,104 +3115,6 @@ static void suppressingFeature(PrintOptions &options, Feature feature,
 #include "swift/Basic/Features.def"
   }
   llvm_unreachable("exhaustive switch");
-}
-
-using BasicFeatureSet = FixedBitSet<numFeatures(), Feature>;
-
-class FeatureSet {
-  BasicFeatureSet required;
-
-  // Stored inverted: index i actually represents
-  // Feature(numFeatures() - i)
-  //
-  // This is the easiest way of letting us iterate from largest to
-  // smallest, i.e. from the newest to the oldest feature, which is
-  // the order in which we need to emit #if clauses.
-  using SuppressibleFeatureSet = FixedBitSet<numFeatures(), size_t>;
-  SuppressibleFeatureSet suppressible;
-
-public:
-  class SuppressibleGenerator {
-    SuppressibleFeatureSet::iterator i, e;
-    friend class FeatureSet;
-    SuppressibleGenerator(const SuppressibleFeatureSet &set)
-      : i(set.begin()), e(set.end()) {}
-
-  public:
-    bool empty() const { return i == e; }
-    Feature next() { return Feature(numFeatures() - *i++); }
-  };
-
-  bool empty() const {
-    return required.empty() && suppressible.empty();
-  }
-
-  bool hasAnyRequired() const {
-    return !required.empty();
-  }
-  const BasicFeatureSet &requiredFeatures() const {
-    return required;
-  }
-
-  bool hasAnySuppressible() const {
-    return !suppressible.empty();
-  }
-  SuppressibleGenerator generateSuppressibleFeatures() const {
-    return SuppressibleGenerator(suppressible);
-  }
-
-  enum InsertOrRemove: bool {
-    Insert = true, Remove = false
-  };
-
-  void collectRequiredFeature(Feature feature, InsertOrRemove operation) {
-    assert(!isSuppressibleFeature(feature));
-    required.insertOrRemove(feature, operation == Insert);
-  }
-
-  void collectSuppressibleFeature(Feature feature, InsertOrRemove operation) {
-    assert(isSuppressibleFeature(feature));
-    suppressible.insertOrRemove(numFeatures() - size_t(feature),
-                                operation == Insert);
-  }
-
-  /// Go through all the features used by the given declaration and
-  /// either add or remove them to this set.
-  void collectFeaturesUsed(Decl *decl, InsertOrRemove operation) {
-    // Go through each of the features, checking whether the
-    // declaration uses that feature.
-#define LANGUAGE_FEATURE(FeatureName, SENumber, Description)                   \
-  if (usesFeature##FeatureName(decl))                                          \
-    collectRequiredFeature(Feature::FeatureName, operation);
-#define SUPPRESSIBLE_LANGUAGE_FEATURE(FeatureName, SENumber, Description)      \
-  if (usesFeature##FeatureName(decl))                                          \
-    collectSuppressibleFeature(Feature::FeatureName, operation);
-#include "swift/Basic/Features.def"
-  }
-};
-
-/// Get the set of features that are uniquely used by this declaration, and are
-/// not part of the enclosing context.
-static FeatureSet getUniqueFeaturesUsed(Decl *decl) {
-  // Add all the features used by this declaration.
-  FeatureSet features;
-  features.collectFeaturesUsed(decl, FeatureSet::Insert);
-
-  // Remove all the features used by all enclosing declarations.
-  Decl *enclosingDecl = decl;
-  while (!features.empty()) {
-    // Find the next outermost enclosing declaration.
-    if (auto accessor = dyn_cast<AccessorDecl>(enclosingDecl))
-      enclosingDecl = accessor->getStorage();
-    else
-      enclosingDecl = enclosingDecl->getDeclContext()->getAsDecl();
-    if (!enclosingDecl)
-      break;
-
-    features.collectFeaturesUsed(enclosingDecl, FeatureSet::Remove);
-  }
-
-  return features;
 }
 
 static void printCompatibilityCheckIf(ASTPrinter &printer, bool isElseIf,
