+++ lib/AST/CMakeLists.txt
@@ -48,6 +48,7 @@ add_swift_host_library(swiftAST STATIC
   ExistentialGeneralization.cpp
   Expr.cpp
   ExtInfo.cpp
+  FeatureSet.cpp
   FineGrainedDependencies.cpp
   FineGrainedDependencyFormat.cpp
   FreestandingMacroExpansion.cpp
